# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EnergieGo API",
        "description": "Documentation OpenAPI pour l'API EnergieGo",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/promotions": {
            "get": {
                "summary": "Lister les promotions",
                "operationId": "85d8d71a0110712ce76708e909112e34",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Promotion"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Créer une promotion",
                "operationId": "4992bbc1b3e1bb4a32e15f94108f3fea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Promotion"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Promotion"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/promotions/{id}": {
            "get": {
                "summary": "Afficher une promotion",
                "operationId": "b4808b9910086dfb24d30e67efe549c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Promotion"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Mettre à jour une promotion",
                "operationId": "8792c4fa3e28dff76f7aaa7760dd81ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Promotion"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Promotion"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Supprimer une promotion",
                "operationId": "93589e73e0f79aac0abdcf09903b06d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                }
            }
        },
        "/api/utilisateurs": {
            "get": {
                "summary": "Lister les utilisateurs",
                "operationId": "27a03cbccef26b92bf6d80712f8b8ab8",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Utilisateur"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Créer un utilisateur",
                "operationId": "dded82ccae378b45b0c84f76ec078823",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Utilisateur"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Utilisateur"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/utilisateurs/{id}": {
            "get": {
                "summary": "Afficher un utilisateur",
                "operationId": "565dd7bc723d9a51b807156bd330fb2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Utilisateur"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "put": {
                "summary": "Mettre à jour un utilisateur",
                "operationId": "08fd8a9313115a94181f3651f05b6fa6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Utilisateur"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Utilisateur"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "delete": {
                "summary": "Supprimer un utilisateur",
                "operationId": "c0fc30a3d234d962f29c73d25304de83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Utilisateur": {
                "required": [
                    "nom",
                    "prenom",
                    "email",
                    "motDePasse",
                    "role",
                    "dateInscription"
                ],
                "properties": {
                    "id_utilisateur": {
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "John"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@exemple.com"
                    },
                    "motDePasse": {
                        "type": "string",
                        "example": "hashed"
                    },
                    "role": {
                        "type": "string",
                        "example": "client"
                    },
                    "dateInscription": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Promotion": {
                "required": [
                    "codePromo",
                    "reduction",
                    "dateDebut",
                    "dateFin"
                ],
                "properties": {
                    "id_promotion": {
                        "type": "integer",
                        "example": 1
                    },
                    "codePromo": {
                        "type": "string"
                    },
                    "reduction": {
                        "type": "number"
                    },
                    "dateDebut": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "dateFin": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    }
}